% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banknote.R
\docType{data}
\name{banknote}
\alias{banknote}
\title{Bank Note Authentication}
\format{
A data frame with 1372 non-null observations of 5 variables.

\code{variance} variance of Wavelet Transform

\code{skewness} skewness of Wavelet Transform

\code{curtosis} curtosis of Wavelet Transform

\code{entropy} entropy of Wavelet Transform

\code{class} forged 1 or geniuine 0
}
\usage{
banknote
}
\description{
Banknote data extracted from images of genuine and forged banknotes.
}
\keyword{datasets}
