% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classifier}
\alias{classifier}
\title{Prediction model}
\usage{
classifier(
  formula,
  df,
  split = c(0.7, 0.2, 0.1),
  method = "all",
  return_all = FALSE
)
}
\arguments{
\item{formula}{Indicates dependent and explanatory variables}

\item{df}{data provided in data.frame format}

\item{split}{proportion size of train, test and check datasets}

\item{method}{indicates which ML method should be used to predict data. Avaiable `logistic regression`, `random forest`, `SVM` and `all`.}

\item{return_all}{boolean, if `TRUE` it will return all datasets with predictions column}
}
\value{
Function returns a list containing accuracy results and data frames with predictions
}
\description{
This function creates and calculates prediction model and checks its accuracy on test model.
}
\examples{
classifier(Survived ~ ., titanic, split = c(0.7, 0.2, 0.1), method = 'svm', return_all = FALSE)
\dontrun{classifier(Survived ~ ., split = c(0.7, 0.2, 0.1), method = 'all', return_all = TRUE)}
}
