% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_df}
\alias{split_df}
\title{Data splitter}
\usage{
split_df(df, prob = c(0.7, 0.2, 0.1))
}
\arguments{
\item{df}{dataset}

\item{prob}{proportion size of train, test and check datasets}
}
\value{
It returns a list with 3 dataframes
}
\description{
This function splits data frame into 3 smaller data frames without replacements.
}
